requirements.txt
fastapi==0.95.0
uvicorn==0.20.0
pydantic==1.10.7
httpx==0.23.3
python-dotenv==0.21.1
beautifulsoup4==4.11.2 